+++ LocalPackages/Client/Sources/Client/Enumerations/FeatureFlagType.swift
@@ -26,7 +26,6 @@ import ProtonCoreFeatureFlags
 /// Example:
 ///    case passSharingV1 = "PassSharingV1"
 public enum FeatureFlagType: String, CaseIterable, FeatureFlagTypeProtocol {
-    case passCustomTypeV1 = "PassCustomTypeV1"
     case passRenameAdminToManager = "PassRenameAdminToManager"
     case passHideShowVault = "PassHideShowVault"
     case passUserEventsV1 = "PassUserEventsV1"