+++ LocalPackages/Client/Sources/Client/Managers/InAppNotificationManager.swift
@@ -28,11 +28,13 @@ public enum InAppNotificationDisplayState: Sendable {
 }
 
 public protocol InAppNotificationManagerProtocol: Sendable {
-    func fetchNotifications(offsetId: String?, reset: Bool) async throws -> [InAppNotification]
-    func getNotificationToDisplay() async throws -> InAppNotification?
+    var notificationToDisplay: CurrentValueSubject<InAppNotification?, Never> { get }
+
+    func refreshNotifications()
     func updateNotificationState(notificationId: String, newState: InAppNotificationState) async throws
     func updateNotificationTime(_ date: Date) async throws
     func updateDisplayState(_ state: InAppNotificationDisplayState) async
+    func updateCurrentPromoState(_ state: InAppNotificationState)
 
     // MARK: - Qa only accessible function to test mock notifications
 
@@ -41,12 +43,6 @@ public protocol InAppNotificationManagerProtocol: Sendable {
 
     // periphery:ignore
     @_spi(Test) func getCurrentNofications() async -> [InAppNotification]
-}
-
-public extension InAppNotificationManagerProtocol {
-    func fetchNotifications(offsetId: String? = nil, reset: Bool = true) async throws -> [InAppNotification] {
-        try await fetchNotifications(offsetId: offsetId, reset: reset)
-    }
 }
 
 public actor InAppNotificationManager: InAppNotificationManagerProtocol {
@@ -62,6 +58,8 @@ public actor InAppNotificationManager: InAppNotificationManagerProtocol {
 
     private var mockNotification: InAppNotification?
 
+    public nonisolated let notificationToDisplay: CurrentValueSubject<InAppNotification?, Never> = .init(nil)
+
     public init(repository: any InAppNotificationRepositoryProtocol,
                 timeDatasource: any LocalNotificationTimeDatasourceProtocol,
                 userManager: any UserManagerProtocol,
@@ -76,8 +74,22 @@ public actor InAppNotificationManager: InAppNotificationManagerProtocol {
 }
 
 public extension InAppNotificationManager {
-    func fetchNotifications(offsetId: String?,
-                            reset: Bool) async throws -> [InAppNotification] {
+    nonisolated func refreshNotifications() {
+        Task { [weak self] in
+            guard let self else { return }
+            let displayState = await displayState
+            guard displayState == .inactive else { return }
+            do {
+                _ = try await fetchNotifications()
+                let notification = try await getNotificationToDisplay()
+                notificationToDisplay.send(notification)
+            } catch {
+                logger.error(error)
+            }
+        }
+    }
+
+    func fetchNotifications(offsetId: String? = nil, reset: Bool = true) async throws -> [InAppNotification] {
         let userId = try await userManager.getActiveUserId()
         let paginatedNotifications = try await repository
             .getPaginatedNotifications(lastNotificationId: offsetId,
@@ -100,24 +112,30 @@ public extension InAppNotificationManager {
         if let mockNotification {
             return mockNotification.canBeDisplayed(timestampDate: timestampDate.toInt) ? mockNotification : nil
         }
-        guard try await shouldDisplayNotification() else {
-            return nil
-        }
-        return notifications.filter { notification in
+
+        guard let notification = notifications.filter({ notification in
             notification.canBeDisplayed(timestampDate: timestampDate.toInt)
-        }.max(by: { lhs, rhs in
+        }).max(by: { lhs, rhs in
             // Priority descending
             if lhs.priority != rhs.priority {
                 return lhs.priority < rhs.priority
             }
             // StartTime ascending
             return lhs.startTime > rhs.startTime
-        })
+        }) else {
+            return nil
+        }
+
+        let shouldDisplay = try await shouldDisplayNotification()
+        guard notification.displayType == .promo || shouldDisplay else {
+            return nil
+        }
+        return notification
     }
 
     func updateNotificationState(notificationId: String, newState: InAppNotificationState) async throws {
         guard mockNotification == nil else {
-            mockNotification?.state = newState.rawValue
+            mockNotification?.state = newState
             return
         }
         let userId = try await userManager.getActiveUserId()
@@ -126,9 +144,10 @@ public extension InAppNotificationManager {
                                                       userId: userId)
         if newState == .dismissed {
             try await repository.remove(notificationId: notificationId, userId: userId)
+            notificationToDisplay.send(nil)
         }
         if let index = notifications.firstIndex(where: { $0.id == notificationId }) {
-            notifications[index].state = newState.rawValue
+            notifications[index].state = newState
         }
     }
 
@@ -139,6 +158,15 @@ public extension InAppNotificationManager {
 
     func updateDisplayState(_ state: InAppNotificationDisplayState) async {
         displayState = state
+    }
+
+    nonisolated func updateCurrentPromoState(_ state: InAppNotificationState) {
+        guard var notification = notificationToDisplay.value,
+              notification.displayType == .promo else {
+            return
+        }
+        notification.state = state
+        notificationToDisplay.send(notification)
     }
 }
 
@@ -176,8 +204,15 @@ private extension InAppNotificationManager {
 
 private extension InAppNotification {
     func canBeDisplayed(timestampDate: Int) -> Bool {
-        !hasBeenRead &&
-            startTime <= timestampDate &&
-            (endTime ?? .max) >= timestampDate
+        let validStartTime = startTime <= timestampDate
+        let validEndTime = (endTime ?? .max) >= timestampDate
+        let validTime = validStartTime && validEndTime
+
+        return switch displayType {
+        case .promo:
+            validTime
+        case .banner, .modal:
+            isUnread && validTime
+        }
     }
 }