+++ iOS/Scenes/Homepage/Create & Edit Items/CreateEditAliasView.swift
@@ -160,16 +160,14 @@ struct CreateEditAliasView: View {
                         senderNameRow
                     }
 
-                    if viewModel.customTypeEnabled {
-                        EditCustomFieldSections(focusedField: $focusedField,
-                                                focusedCustomField: viewModel.recentlyAddedOrEditedField,
-                                                contentType: .alias,
-                                                fields: $viewModel.customFields,
-                                                canAddMore: viewModel.canAddMoreCustomFields,
-                                                onAddMore: { viewModel.requestAddCustomField(to: nil) },
-                                                onEditTitle: viewModel.requestEditCustomFieldTitle,
-                                                onUpgrade: { viewModel.upgrade() })
-                    }
+                    EditCustomFieldSections(focusedField: $focusedField,
+                                            focusedCustomField: viewModel.recentlyAddedOrEditedField,
+                                            contentType: .alias,
+                                            fields: $viewModel.customFields,
+                                            canAddMore: viewModel.canAddMoreCustomFields,
+                                            onAddMore: { viewModel.requestAddCustomField(to: nil) },
+                                            onEditTitle: viewModel.requestEditCustomFieldTitle,
+                                            onUpgrade: { viewModel.upgrade() })
 
                     FileAttachmentsEditSection(files: viewModel.fileUiModels,
                                                isFetching: viewModel.isFetchingAttachedFiles,