+++ CHANGELOG.md
@@ -1,8 +1,11 @@
-# Version x.x.x
+# Version 1.17.9
 - Fixed password generator partially hidden behind the keyboard on iOS 26
+- Show a warning when creating or editing items in shared vaults
 
 Others:
-- Removed feature discovery for custom items
+- Removed feature discovery & flag for custom items
+- Added support for promo in-app notifications
+- Updated Rust library from 0.19.0 to 0.26.0
 
 # Version 1.17.8
 Fixed un-tappable create item button in the AutoFill extension on iOS 26